backend/requirements.txt
fastapi[all]
sqlmodel
oracledb
uvicorn
passlib
python-jose
bcrypt
pyjwt